// src/domain/entities/auth/forgot-password.interface.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rent System API
 * Rent System API Description
 *
 * The version of the OpenAPI document: 1.0.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface ForgotPasswordPatchReq
 */
export interface ForgotPasswordPatchReq {
  /**
   *
   * @type {string}
   * @memberof ForgotPasswordPatchReq
   */
  phone_number: string
  /**
   *
   * @type {string}
   * @memberof ForgotPasswordPatchReq
   */
  password: string
  /**
   *
   * @type {string}
   * @memberof VerifyOtpPostReq
   */
  otp: string
  /**
   *
   * @type {string}
   * @memberof VerifyOtpPostReq
   */
  device_id: string
  /**
   *
   * @type {string}
   * @memberof VerifyOtpPostReq
   */
  updated_by?: string
  /**
   *
   * @type {Date}
   * @memberof VerifyOtpPostReq
   */
  updated_at: Date
}

export interface ForgotPasswordUpdateDetailsReq {
  /**
   *
   * @type {string}
   * @memberof ForgotPasswordUpdateDetailsReq
   */
  phone_number: string
  /**
   *
   * @type {string}
   * @memberof ForgotPasswordUpdateDetailsReq
   */
  password: string
  /**
   *
   * @type {Date}
   * @memberof ForgotPasswordUpdateDetailsReq
   */
  updated_at?: Date
}